Crea WorldmapActor
En vez de tener todo en el WorldmapStage.
