+++ core/src/main/screens/running/CountrySelectionListener.kt
@@ -4,8 +4,6 @@ import Country
 
 interface CountrySelectionListener {
     fun onCountrySelected(country: Country)
-}
-
-class NoCountrySelectionListener: CountrySelectionListener {
-    override fun onCountrySelected(country: Country) {}
+    fun onCountryMouseOver(country: Country)
+    fun onCountryExit(country: Country)
 }
